Fix for [HER-1627] incorrect source tag in de-duped crawls * AList.java, HashtableAList.java     new copyKeysFrom() with boolean attribute "clobber" * PersistLoadProcessor.java     innerProcess() do not clobber when loading persisted info

+++ heritrix/src/java/st/ata/util/HashtableAList.java
@@ -140,11 +140,17 @@ public class HashtableAList implements MutableAList, Serializable {
     }
     
     public void copyKeysFrom(Iterator keys, AList other) {
+        copyKeysFrom(keys, other, true);  // clobber
+    }
+    
+    public void copyKeysFrom(Iterator keys, AList other, boolean clobber) {
         for (; keys.hasNext();) {
             String key = (String)keys.next();
+            if (!clobber && containsKey(key))
+                continue;
             Object value = other.getObject(key);
             // TODO: consider shallow or deep copy in some cases?
-            // perhaps controlled by a third parameter?
+            // perhaps controlled by an additional parameter?
             if(value!=null) {
                 putObject(key,value);
             }
